FM : Don't register controls if service not running
Issue : Lockscreen controls, pressing play when nothing is active tried to
turn on fm radio. We only need to control the radio when the service is active.

I tried to address this by putting a check on the abortbroadcast call in
FMMediaButtonIntentReceiver but even when abortBroadcast is not called, none of the
other receivers get the broadcast (Even though it is a ordered broadcast).

So to address this issue, i just moved the class to an inner class and removed the declaration
from the manifest. Hence the buttons are only registered when service is active.

Change-Id: Iea7f8f87fed7a33f0bfa06d83da3a54f4c208c1e