FM : Unregister/Register based on service status
Make it so :
service is on -> Register
service is off -> Unregister

This avoids problems like turning off fm via toggle and pressing home rather than back
The controls remain registered and lockscreen controls attempt to start fm radio.

Move media button receiver to an inner class in the service.

This is to prevent media button broadcasts from being received if service is not active.

Change-Id: Iea7f8f87fed7a33f0bfa06d83da3a54f4c208c1e