FM : Fixes
This patchset addresses 3 issues :

1. If no media service is active/registered. Lockscreen play attempts to start,
the service. This patch makes it so the receiver is only enabled when service is active.

2. Lockscreen station frequency doesnt update if not ui is not binded to service.
This patch makes a seek update the lockscreen info even if ui is not binded to service.

3. Restarting the activity multiple times causes static noise upon launch. This patch,
addresses that issue by checking if the fm is on, if it is avoid unecessary settings from
being re-applied.

Patchset 7 : Fix seekbar not being updated.
             Change radio station only if different from currently set
	     Set mMuted to false in fmOn()

Change-Id: Iea7f8f87fed7a33f0bfa06d83da3a54f4c208c1e